BLZT-23: impl Immutable annotation for entity that skip first level c… (#24)
* BLZT-23: impl Immutable annotation for entity that skip first level cache and use just as dictionary only select (update, delete not allowed)

* BLZT-25: add support simple HQL for repository.

* BLZT-7: improve pars logic to support findByAgeLessThan or findByAgeGreaterThanEqual etc.

* BLZT-7: rename some fields.

---------

+++ core/src/main/java/io/github/blyznytsiaorg/bibernate/session/DefaultBibernateSession.java
@@ -2,11 +2,14 @@ package io.github.blyznytsiaorg.bibernate.session;
 
 import io.github.blyznytsiaorg.bibernate.dao.Dao;
 import io.github.blyznytsiaorg.bibernate.exception.BibernateGeneralException;
+import io.github.blyznytsiaorg.bibernate.exception.ImmutableEntityException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 import java.util.List;
 import java.util.Optional;
+
+import static io.github.blyznytsiaorg.bibernate.utils.EntityReflectionUtils.isImmutable;
 
 /**
  *
@@ -17,6 +20,7 @@ import java.util.Optional;
 @Slf4j
 public class DefaultBibernateSession implements BibernateSession {
 
+    public static final String IMMUTABLE_ENTITY_S_NOT_ALLOWED_TO_CHANGE = "Immutable entity %s not allowed to change";
     private final Dao dao;
     private boolean closed;
 
@@ -33,9 +37,15 @@ public class DefaultBibernateSession implements BibernateSession {
     }
 
     @Override
-    public <T> List<T> findBy(Class<T> entityClass, String whereQuery, Object[] bindValues) {
+    public <T> List<T> findByWhere(Class<T> entityClass, String whereQuery, Object[] bindValues) {
         verifySessionNotClosed();
-        return dao.findBy(entityClass, whereQuery, bindValues);
+        return dao.findByWhere(entityClass, whereQuery, bindValues);
+    }
+
+    @Override
+    public <T> List<T> findByQuery(Class<T> entityClass, String query, Object[] bindValues) {
+        verifySessionNotClosed();
+        return dao.findByQuery(entityClass, query, bindValues);
     }
 
     @Override
@@ -47,19 +57,28 @@ public class DefaultBibernateSession implements BibernateSession {
     @Override
     public <T> T save(Class<T> entityClass, Object entity) {
         verifySessionNotClosed();
+        if (isImmutable(entityClass)) {
+            throw new ImmutableEntityException(
+                    IMMUTABLE_ENTITY_S_NOT_ALLOWED_TO_CHANGE.formatted(entityClass)
+            );
+        }
         return dao.save(entityClass, entity);
     }
 
     @Override
     public <T> void delete(Class<T> entityClass, Object primaryKey) {
         verifySessionNotClosed();
+        if (isImmutable(entityClass)) {
+            log.warn(IMMUTABLE_ENTITY_S_NOT_ALLOWED_TO_CHANGE.formatted(entityClass));
+            return;
+        }
         dao.delete(entityClass, primaryKey);
     }
 
     @Override
     public void close() {
         closed = true;
-        log.info("Close session...");
+        log.trace("Close session...");
     }
 
     @Override
